[Chore] #159 - 한 마디 텍스트뷰 컴포넌트화 #1
